requirements.txt
Flask
python-binance
flask